Make `Open` and `Close` functions lowercase
To comply with the standard Python implementation of other libraries
that open compressed file types (i.e. `gzip`, `bz2`), rename `Open` and
`Close` to `open` and `close` respectively.

These functions were previously implemented with capital letters to be
consistent with the implementation of the other exported classes in
`Interface.cpp`, but downstream code that may rely on `DsrcReadInMemory`
implementations may implicitly assume `open` and `close` were
named in the same manner as `gzip.open`/`gzip.close` are