Implement correct logic for `readline`
Utilize a the C++ standard implementation for a `Queue` to store lines
from a retrieved `chunk`

If the `Queue` is empty, the next chunk needs to be retrieved. This
needs to be repeated until no more chunks can be retrieved from the
opened `.dsrc` file.

If the `Queue` is not empty, then the next line in the `Queue` needs to
be popped out of the `Queue` and returned